+++ packages/components/src/Copy/index.ts
@@ -1,0 +1,2 @@
+import Copy from "./index.vue"
+export { Copy }