+++ packages/components/src/DialogPanel/index.ts
@@ -1,0 +1,2 @@
+import DialogPanel from "./index.vue"
+export { DialogPanel }