+++ packages/components/src/Spinner/index.ts
@@ -1,0 +1,2 @@
+import Spinner from "./index.vue"
+export { Spinner }