+++ packages/components/src/ThreePanel/three/useCollider.ts
@@ -9,7 +9,7 @@ export function useCollider(bot: Bot) {
   const tempVector = new THREE.Vector3()
   const tempVector2 = new THREE.Vector3()
   const tempBox = new THREE.Box3()
-  const tempBoxHelper = new THREE.Box3Helper(tempBox, 0xff0000) // 边界框颜色为红色
+  const tempBoxHelper = new THREE.Box3Helper(tempBox, 0xff0000)
   const tempMat = new THREE.Matrix4()
   const tempSegment = new THREE.Line3()
   const upVector = new THREE.Vector3(0, 1, 0)
@@ -34,16 +34,6 @@ export function useCollider(bot: Bot) {
     }
   }
   function init() {
-    // 创建几何体
-    // const points = [capsule.segment.start, capsule.segment.end]
-    // const geometry = new THREE.BufferGeometry().setFromPoints(points)
-    // const material = new THREE.LineBasicMaterial({
-    //   color: 0xff0000,
-    //   linewidth: 2,
-    // })
-    // const line = new THREE.Line(geometry, material)
-    // bot.core.scene.add(line)
-
     bot.core.scene.add(tempBoxHelper)
     player = new THREE.Mesh(new RoundedBoxGeometry(1.0, 2.0, 1.0, 10, 0.5), new THREE.MeshStandardMaterial())
     player.material.wireframe = true
@@ -89,7 +79,7 @@ export function useCollider(bot: Bot) {
       player.position.addScaledVector(tempVector, bot.move.speed * delta)
     }
     if (bot.move.forward || bot.move.backward || bot.move.right || bot.move.left) {
-      // 计算人物相对于相机的相对旋转角度
+      // calculate the relative rotation angle of the character relative to the camera
       const forward = bot.move.forward ? 1 : bot.move.backward ? -1 : 0
       const right = bot.move.right ? 1 : bot.move.left ? -1 : 0
       // const botRotation =
@@ -104,7 +94,7 @@ export function useCollider(bot: Bot) {
     player.updateMatrixWorld()
     // adjust player position based on collisions
     tempBox.makeEmpty()
-    // 获取bvh的世界矩阵并计算其逆矩阵，逆矩阵用于从世界坐标转换回局部坐标
+    // get bvh world matrix and calculate its inverse matrix, which is used to convert from world to local coordinates
     tempMat.copy(collider.matrixWorld).invert()
     tempSegment.copy(capsule.segment)
     // get the position of the capsule in the local space of the collider
@@ -169,25 +159,6 @@ export function useCollider(bot: Bot) {
       reset()
     }
   }
-  /**
-   * 相机碰撞处理
-   * TODO: 未生效
-   */
-  // private cameraCollision(bvh: THREE.Object3D[]) {
-  //   if (!this.model) return
-  //   const ray_direction = new THREE.Vector3()
-  //   ray_direction.subVectors(this.core.camera.position, this.model.scene.position).normalize()
-  //   // 设置镜头光线
-  //   this.cameraRaycaster.set(this.model.scene.position, ray_direction)
-  //   const intersects = this.cameraRaycaster.intersectObjects(bvh)
-  //   if (intersects.length) {
-  //     // 找到碰撞点后还需要往前偏移一点，不然还是可能会看到穿模
-  //     const offset = new THREE.Vector3() // 定义一个向前移动的偏移量
-  //     offset.copy(ray_direction).multiplyScalar(-0.5) // 计算偏移量，这里的distance是想要向前移动的距离
-  //     const new_position = new THREE.Vector3().addVectors(intersects[0].point, offset) // 计算新的相机位置
-  //     this.core.camera.position.copy(new_position)
-  //   }
-  // }
   function teleport(x: number, y: number, z: number) {
     player.position.set(x, y + 1.5, z)
   }
@@ -195,7 +166,7 @@ export function useCollider(bot: Bot) {
     return player
   }
   /**
-   * 预设配置
+   * reset config
    */
   function reset() {
     velocity.set(0, 0, 0)
@@ -247,72 +218,3 @@ export function useCollider(bot: Bot) {
     setCapsule,
   }
 }
-// export function useRay(bot: Bot) {
-//   const botRaycaster: THREE.Raycaster = new THREE.Raycaster()
-//   let rayLine: THREE.Line<
-//     THREE.BufferGeometry<THREE.NormalBufferAttributes>,
-//     THREE.LineBasicMaterial,
-//     THREE.Object3DEventMap
-//   >
-//   function getRayLine(raycaster: THREE.Raycaster) {
-//     const ray = raycaster.ray
-//     const lineMaterial = new THREE.LineBasicMaterial({ color: 0xff0000 })
-//     const lineGeometry = new THREE.BufferGeometry().setFromPoints([
-//       ray.origin,
-//       new THREE.Vector3(
-//         ray.origin.x + ray.direction.x * 10,
-//         ray.origin.y + ray.direction.y * 10,
-//         ray.origin.z + ray.direction.z * 10
-//       ),
-//     ])
-
-//     const line = new THREE.Line(lineGeometry, lineMaterial)
-//     return line
-//   }
-//   function update(_delta: number) {
-//     if (!bot.model?.scene) return
-//     let origin = bot.model.scene.position.clone().add(new THREE.Vector3(0, 1, 0))
-//     let direction = new THREE.Vector3()
-//     bot.model.scene.getWorldDirection(direction)
-//     direction.normalize()
-//     botRaycaster.ray.origin.copy(origin)
-//     botRaycaster.ray.direction.copy(direction)
-//     const rayEnd = new THREE.Vector3(
-//       botRaycaster.ray.origin.x + botRaycaster.ray.direction.x * 10,
-//       botRaycaster.ray.origin.y + botRaycaster.ray.direction.y * 10,
-//       botRaycaster.ray.origin.z + botRaycaster.ray.direction.z * 10
-//     )
-//     rayLine.geometry.setAttribute(
-//       "position",
-//       new THREE.Float32BufferAttribute(
-//         [botRaycaster.ray.origin.x, botRaycaster.ray.origin.y, botRaycaster.ray.origin.z, rayEnd.x, rayEnd.y, rayEnd.z],
-//         3
-//       )
-//     )
-//     rayLine.geometry.computeBoundingSphere()
-//   }
-//   rayLine = getRayLine(botRaycaster)
-//   bot.core.scene.add(rayLine)
-//   return { update }
-// }
-
-// export function useBotBvh(bot: Bot): THREE.Mesh | undefined {
-//   if (!bot.model?.scene) return
-//   // 构建 bvh
-//   const staticGenerator = new StaticGeometryGenerator(bot.model.scene)
-//   staticGenerator.attributes = ["position"]
-//   const generateGeometry = staticGenerator.generate() as THREE.BufferGeometry<THREE.NormalBufferAttributes> & {
-//     boundsTree: MeshBVH
-//   }
-//   generateGeometry.boundsTree = new MeshBVH(generateGeometry, { lazyGeneration: false } as MeshBVHOptions)
-//   bot.bvh = new THREE.Mesh(generateGeometry)
-//   ;(bot.bvh.material as THREE.MeshBasicMaterial).wireframe = true
-//   // ;(bot.bvh.material as THREE.MeshBasicMaterial).opacity = 0.5
-//   // ;(bot.bvh.material as THREE.MeshBasicMaterial).transparent = true
-//   bot.bvhHelper = new MeshBVHHelper(bot.bvh, 10)
-//   bot.bvh.visible = true
-//   bot.bvhHelper.visible = true
-//   // bot.core.scene.add(bot.bvhHelper)
-//   // bot.core.scene.add(bot.bvh)
-//   console.log(bot.bvh, bot.bvhHelper)
-// }