+++ packages/components/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @toolmain/components
 
+## 1.2.4
+
+### Patch Changes
+
+- fix: fix some problems
+
+  feat: add new components
+
+- Updated dependencies
+  - @toolmain/shared@1.2.3
+  - @toolmain/libs@1.1.3
+
+## 1.2.3
+
+### Patch Changes
+
+- feat: add new components
+
 ## 1.2.2
 
 ### Patch Changes